unknown experiment

+++ app/models/schedule.rb
@@ -10,23 +10,23 @@ class Schedule < ActiveRecord::Base
 	# this method accepts the yaml contents of a schedule file
 	#
 	def load(contents)
-
-		# save the NAME of the current schedule item, to restore later
-		backup_position = current.name if current
+		if contents
+			# save the NAME of the current schedule item, to restore later
+			backup_position = current.name if current
 		
-		# delete al items
-		schedule_spans.delete_all
+			# delete al items
+			schedule_spans.delete_all
 		
-		# create all items
-		contents.each do |name, items|
-			span = schedule_spans.where(name: name).first_or_initialize
-			span.content = items.to_yaml
-			span.save
+			# create all items
+			contents.each do |name, items|
+				span = schedule_spans.where(name: name).first_or_initialize
+				span.content = items.to_yaml
+				span.save
+			end
+		
+			# restore 'current' item
+			update_attribute(:current, backup_position && self.schedule_spans.find_by_name(backup_position))
 		end
-		
-		# restore 'current' item
-		update_attribute(:current, backup_position && self.schedule_spans.find_by_name(backup_position))
-		
 	end
 
 end
